requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
openai==1.51.0          # or anthropic==0.34.2
langchain==0.3.3
langchain-openai==0.2.1 # if using LangChain + OpenAI
pydantic>=2.12.0